# openapi/definicion.yaml
openapi: 3.1.0
info:
  title: CatalogIA API
  description: FastAPI backend for CatalogIA platform with AI recommendations
  version: 1.0.0
  contact:
    name: Equipo 22
    email: 'equipo22@tec.mx'
    url: http://localhost:8000
servers:
- url: http://localhost:8000
  description: Local development server
paths:
...
  /api/v1/orders/:
    post:
      tags:
      - orders
      summary: Create Order
      description: "Crear una nueva orden de compra.\n\nArgs:\n    `db`: Sesión de\
        \ base de datos\n    `order_in`: Datos de la orden a crear\n    `current_user`:\
        \ Usuario autenticado que realiza la orden\n\nReturns:\n    `OrderResponse`:\
        \ Orden creada con el total calculado\n    \nRaises:\n    `HTTPException`:\
        \ 404 si el producto no existe\n    `HTTPException`: 400 si no hay suficiente\
        \ stock"
      operationId: create_order_api_v1_orders__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - orders
      summary: List User Orders
      description: "Obtener todas las órdenes del usuario autenticado.\n\nArgs:\n\
        \    `db`: Sesión de base de datos\n    `current_user`: Usuario autenticado\
        \ actual\n    `skip`: Número de órdenes a omitir (para paginación)\n    `limit`:\
        \ Número máximo de órdenes a retornar\n\nReturns:\n    `OrderList`: Lista\
        \ paginada de órdenes del usuario"
      operationId: list_user_orders_api_v1_orders__get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Número de órdenes a saltar para paginación
          default: 0
          title: Skip
        description: Número de órdenes a saltar para paginación
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Límite de órdenes por página
          default: 100
          title: Limit
        description: Límite de órdenes por página
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}:
    get:
      tags:
      - orders
      summary: Get Order
      description: "Obtener una orden específica por su ID.\n\nArgs:\n    `order_id`:\
        \ UUID de la orden a buscar\n    `db`: Sesión de base de datos\n    `current_user`:\
        \ Usuario autenticado actual\n\nReturns:\n    `OrderResponse`: Datos completos\
        \ de la orden\n    \nRaises:\n    `HTTPException`: 404 si la orden no existe\n\
        \    `HTTPException`: 403 si la orden no pertenece al usuario"
      operationId: get_order_api_v1_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - orders
      summary: Cancel Order
      description: "Cancelar una orden (cambiar estado a 'cancelled').\n\nArgs:\n\
        \    `db`: Sesión de base de datos\n    `order_id`: UUID de la orden a cancelar\n\
        \    `current_user`: Usuario autenticado actual\n\nReturns:\n    `dict`: Mensaje\
        \ de confirmación de cancelación\n    \nRaises:\n    `HTTPException`: 404\
        \ si la orden no existe\n    `HTTPException`: 403 si la orden no pertenece\
        \ al usuario\n    `HTTPException`: 400 si la orden ya está completada"
      operationId: cancel_order_api_v1_orders__order_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}/status:
    put:
      tags:
      - orders
      summary: Update Order Status
      description: "Actualizar el estado de una orden.\n\nArgs:\n    `db`: Sesión\
        \ de base de datos\n    `order_id`: UUID de la orden a actualizar\n    `status_update`:\
        \ Nuevo estado de la orden\n    `current_user`: Usuario autenticado actual\n\
        \nReturns:\n    `OrderResponse`: Orden con el estado actualizado\n    \nRaises:\n\
        \    `HTTPException`: 404 si la orden no existe\n    `HTTPException`: 403\
        \ si la orden no pertenece al usuario\n    `HTTPException`: 400 si el estado\
        \ es inválido"
      operationId: update_order_status_api_v1_orders__order_id__status_put
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
      description: Error de validación HTTP que contiene una lista de errores de validación
        detallados.
    OrderCreate:
      properties:
        product_id:
          type: integer
          title: Product Id
        quantity:
          type: integer
          exclusiveMinimum: 0
          title: Quantity
      type: object
      required:
      - product_id
      - quantity
      title: OrderCreate
      description: Esquema para crear una nueva orden de compra. Requiere el ID del
        producto y la cantidad deseada. El total se calcula automáticamente.
    OrderList:
      properties:
        orders:
          items:
            $ref: '#/components/schemas/OrderResponse'
          type: array
          title: Orders
        total:
          type: integer
          title: Total
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - orders
      - total
      - skip
      - limit
      title: OrderList
      description: Respuesta paginada con lista de órdenes del usuario, incluyendo
        el total de órdenes y los parámetros de paginación.
    OrderResponse:
      properties:
        product_id:
          type: integer
          title: Product Id
        quantity:
          type: integer
          exclusiveMinimum: 0
          title: Quantity
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        total_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Amount
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - product_id
      - quantity
      - id
      - user_id
      - total_amount
      - status
      - created_at
      title: OrderResponse
      description: Esquema de respuesta con los detalles completos de una orden, incluyendo
        ID, usuario, producto, cantidad, total calculado, estado y fecha de creación.
    OrderStatusUpdate:
      properties:
        status:
          type: string
          pattern: ^(pending|completed|cancelled)$
          title: Status
      type: object
      required:
      - status
      title: OrderStatusUpdate
      description: 'Esquema para actualizar el estado de una orden. Estados válidos:
        pending (pendiente), completed (completada), cancelled (cancelada).'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
      description: Detalle de un error de validación individual que incluye la ubicación
        del error, el mensaje descriptivo y el tipo de error.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer